+++ CHANGELOG.md
@@ -17,11 +17,12 @@
 * Add support for administrative history, biographical data, physical description, accompanying material and award notes. Refs [UILD-569].
 * Address accessibility issues. Refs [UILD-579].
 * Add support for serials. Refs [UILD-594].
-* Add support for updated extent handling. Rfs [UILD-557].
+* Add support for updated extent handling. Refs [UILD-557].
 * Added Profile selection when the user creates a new Instance. Refs[UILD-573].
 * Add support for dates of publication note (MARC 362). Refs [UILD-606].
 * Add support for book format. Refs [UILD-607].
 * Add support for supplementary content. Refs [UILD-611].
+* Added language subclass support.  Refs [UILD-564].
 
 [UILD-552]:https://folio-org.atlassian.net/browse/UILD-552
 [UILD-544]:https://folio-org.atlassian.net/browse/UILD-544
@@ -45,6 +46,7 @@
 [UILD-606]:https://folio-org.atlassian.net/browse/UILD-606
 [UILD-607]:https://folio-org.atlassian.net/browse/UILD-607
 [UILD-611]:https://folio-org.atlassian.net/browse/UILD-611
+[UILD-564]:https://folio-org.atlassian.net/browse/UILD-564
 
 ## 1.0.5 (2025-04-30)
 * Fixed incorrect behavior when navigating between duplicated resources. Fixes [UILD-553].